Refactor useEffects, add dependecyises and Loading for client secret
