+++ app/app/components/PaymentPage/PaymentPageComponents/Skeletons/TourCardSkeleton.jsx
